Initial formation of the project and some data display
Downloaded the data with the images and CSV files. Unfortunately I didn't push in the commit the actual data (CSV files and the images) because of 100MB cap for commit push.
I Also Started to work on the main Jupiter file, and write a cell which present some data stats.